Return back to intrinsic definition rather than macro, test  operator, use  as LLVM  type representation

+++ test/opencl_tests/test_operations/test_convert/source.cl
@@ -1,4 +1,4 @@
-__kernel void kernel_main(__global int *a, __global const float *b) {
+__kernel void kernel_main(__global int *a, __global const uint *b) {
     int i = get_global_id(0);
     vstore4(convert_int4(vload4(i, b)), i, a);
 }
